requirements.txt
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
filelock==3.4.2
Flask==2.0.2
gdown==4.1.1
google-api-core==1.31.5
google-api-python-client==2.33.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-storage==1.41.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
huggingface-hub==0.4.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
Naked==0.1.31
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyflakes==2.4.0
pymongo==3.12.1
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
sacremoses==0.0.47
scipy==1.7.3
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
tokenizers==0.10.3
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.15.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
